pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-stack-doctor"
version = "0.1.0"
description = "Stack Doctor: an RL environment for diagnosing inference-stack incidents"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "stack_doctor.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["stack_doctor", "stack_doctor.server"]
package-dir = { "stack_doctor" = ".", "stack_doctor.server" = "server" }
